+++ .gitignore
@@ -3,5 +3,5 @@ out
 dist
 .cache
 config.gypi
-src/lib/webui/app
+src/late/webui/app
 src/icons/*.js
