+++ package.json
@@ -11,12 +11,12 @@
     "test": "cross-env NODE_ENV=test mocha",
     "postinstall": "run-s install-app-deps build:webui",
     "install-app-deps": "electron-builder install-app-deps",
-    "clean:webui": "shx rm -rf src/lib/webui/app/",
+    "clean:webui": "shx rm -rf src/late/webui/app/",
     "build": "run-s build:*",
     "build:webui": "run-s build:webui:*",
-    "build:webui:download": "npx ipfs-or-gateway -c QmRvPGBVDBEUvrXrinMKgh8Vu8mVBbTZk3C1jJGfgWchSP -p src/lib/webui/app/",
-    "build:webui:minimize": "shx rm -rf src/lib/webui/app/static/js/*.map && shx rm -rf src/lib/webui/app/static/css/*.map",
-    "build:babel": "babel src --ignore src/lib/webui/app --out-dir out --copy-files",
+    "build:webui:download": "npx ipfs-or-gateway -c QmRvPGBVDBEUvrXrinMKgh8Vu8mVBbTZk3C1jJGfgWchSP -p src/late/webui/app/",
+    "build:webui:minimize": "shx rm -rf src/late/webui/app/static/js/*.map && shx rm -rf src/late/webui/app/static/css/*.map",
+    "build:babel": "babel src --ignore src/late/webui/app --out-dir out --copy-files",
     "build:binaries": "electron-builder --publish onTag"
   },
   "pre-commit": [
