+++ src/late/auto-launch.js
