+++ src/late/index.js
@@ -7,13 +7,11 @@ import downloadHash from './download-hash'
 import takeScreenshot from './take-screenshot'
 import appMenu from './app-menu'
 import addToIpfs from './add-to-ipfs'
-import protocolHandlers from './protocol-handlers'
 import autoUpdater from './auto-updater'
 import tray from './tray'
 
-export default async function () {
-  let ctx = {}
-  await appMenu()
+export default async function (ctx) {
+  await appMenu(ctx)
   await openExternal(ctx)
   await autoUpdater(ctx) // ctx.checkForUpdates
   await registerWebUI(ctx) // ctx.webui, ctx.launchWebUI
@@ -21,7 +19,6 @@ export default async function () {
   await registerDaemon(ctx) // ctx.getIpfsd
   await languageSelector(ctx)
   await addToIpfs(ctx)
-  await protocolHandlers(ctx)
   await autoLaunch(ctx)
   await downloadHash(ctx)
   await takeScreenshot(ctx)
