+++ src/late/language-selector.js
@@ -1,7 +1,7 @@
 import { ipcMain } from 'electron'
 import { i18n, store } from '../utils'
 
-export default async function (ctx) {
+export default function () {
   ipcMain.on('updateLanguage', async (_, lang) => {
     if (lang === store.get('language')) {
       return
