+++ src/late/open-external.js
