+++ src/late/register-daemon.js
