+++ src/late/webui/screenshot.js
