+++ src/utils/index.js
@@ -5,6 +5,7 @@ import logger from './logger'
 import i18n from './i18n'
 import { notify, notifyError } from './notify'
 import quitAndInstall from './quit-and-install'
+import addToIpfs from './add-to-ipfs'
 
 export {
   createDaemon,
@@ -14,5 +15,6 @@ export {
   notify,
   notifyError,
   showErrorMessage,
-  quitAndInstall
+  quitAndInstall,
+  addToIpfs
 }
